opt paywall for non-appstore version
