+++ ExcalidrawZ/ContentView.swift
@@ -39,7 +39,42 @@ struct ContentView: View {
     @State private var isFirstImporting: Bool?
     @State private var cloudContainerEventChangeListener: AnyCancellable?
     
+    @State private var isFirstAppear = true
+
+    
     var body: some View {
+        content()
+            .navigationTitle("")
+            .modifier(PrintModifier())
+            .modifier(WhatsNewSheetViewModifier())
+            .modifier(NewRoomModifier())
+            .modifier(PaywallModifier())
+            .modifier(SearchableModifier())
+            .handlesExternalEvents(preferring: ["*"], allowing: ["*"])
+            .modifier(OpenURLModifier())
+            .modifier(UserActivityHandlerModifier())
+#if os(iOS)
+            .modifier(ApplePencilToolbarModifier())
+#endif
+            .environmentObject(fileState)
+            .environmentObject(exportState)
+            .environmentObject(toolState)
+            .environmentObject(layoutState)
+            .swiftyAlert(logs: true)
+            .bindWindow($window)
+            .containerSizeClassInjection()
+            .onReceive(NotificationCenter.default.publisher(for: .shouldHandleImport)) { notification in
+                handleImport(notification)
+            }
+            .onReceive(NotificationCenter.default.publisher(for: .didOpenFromUrls)) { notification in
+                handleOpenFromURLs(notification)
+            }
+            .withContainerSize()
+            .task { await prepare() }
+    }
+    
+    @MainActor @ViewBuilder
+    private func content() -> some View {
         ZStack {
             if isFirstImporting == nil {
                 Color.clear
@@ -52,13 +87,13 @@ struct ContentView: View {
                 }
             } else {
                 if #available(macOS 14.0, iOS 17.0, *), appPreference.inspectorLayout == .sidebar {
-                    content()
+                    contentView()
                         .inspector(isPresented: $layoutState.isInspectorPresented) {
                             LibraryView()
                                 .inspectorColumnWidth(min: 240, ideal: 250, max: 300)
                         }
                 } else {
-                    content()
+                    contentView()
                     if appPreference.inspectorLayout == .floatingBar {
                         HStack {
                             Spacer()
@@ -82,93 +117,18 @@ struct ContentView: View {
                 }
             }
         }
-        .navigationTitle("")
-        .modifier(PrintModifier())
-        .modifier(WhatsNewSheetViewModifier())
-        .modifier(NewRoomModifier())
-        .modifier(PaywallModifier())
-        .modifier(SearchableModifier())
-        .handlesExternalEvents(preferring: ["*"], allowing: ["*"])
-        .modifier(OpenURLModifier())
-        .modifier(UserActivityHandlerModifier())
-#if os(iOS)
-        .modifier(ApplePencilToolbarModifier())
-#endif
-        .environmentObject(fileState)
-        .environmentObject(exportState)
-        .environmentObject(toolState)
-        .environmentObject(layoutState)
-        .swiftyAlert(logs: true)
-        .bindWindow($window)
-        .containerSizeClassInjection()
-        .onReceive(NotificationCenter.default.publisher(for: .shouldHandleImport)) { notification in
-            guard let urls = notification.object as? [URL] else { return }
-            if window?.isKeyWindow == true {
-                Task.detached {
-                    do {
-                        try await fileState.importFiles(urls)
-                    } catch {
-                        print(error)
-                        await alertToast(error)
-                    }
-                }
-            }
-        }
-        .onReceive(NotificationCenter.default.publisher(for: .didOpenFromUrls)) { notification in
-            if let urls = notification.object as? [URL] {
-                fileState.temporaryFiles.append(contentsOf: urls)
-                fileState.temporaryFiles = Array(Set(fileState.temporaryFiles))
-                if !fileState.isTemporaryGroupSelected || fileState.currentTemporaryFile == nil {
-                    fileState.isTemporaryGroupSelected = true
-                    fileState.currentTemporaryFile = fileState.temporaryFiles.first
-                }
-            }
-        }
-        .withContainerSize()
-        // Check if it is first launch by checking the files count.
-        .task {
-            do {
-                let isEmpty = try viewContext.fetch(NSFetchRequest<File>(entityName: "File")).isEmpty
-                isFirstImporting = isEmpty
-                if isFirstImporting == true, try await CKContainer.default().accountStatus() != .available {
-                    isFirstImporting = false
-                    return
-                } else if !isEmpty {
-                    try await fileState.mergeDefaultGroupAndTrashIfNeeded(context: viewContext)
-                }
-            } catch {
-                alertToast(error)
-            }
-            
-            self.cloudContainerEventChangeListener?.cancel()
-            self.cloudContainerEventChangeListener = NotificationCenter.default.publisher(
-                for: NSPersistentCloudKitContainer.eventChangedNotification
-            ).sink { notification in
-                if let userInfo = notification.userInfo {
-                    if let event = userInfo["event"] as? NSPersistentCloudKitContainer.Event {
-                        // On macOS, event.type will be `setup` only when first launched.
-                        // On iOS, event.type will be `setup` every time it has been launched.
-                        print("NSPersistentCloudKitContainer.eventChangedNotification: \(event.type), succeeded: \(event.succeeded)")
-                        if event.type == .import, event.succeeded, isFirstImporting == true {
-                            isFirstImporting = false
-                            Task { @MainActor in
-                                do {
-                                    try? await Task.sleep(nanoseconds: UInt64(2 * 1e+9))
-                                    try await fileState.mergeDefaultGroupAndTrashIfNeeded(context: viewContext)
-                                } catch {
-                                    alertToast(error)
-                                }
-                            }
-                            self.cloudContainerEventChangeListener?.cancel()
-                        }
-                    }
+        .onAppear {
+            defer { isFirstAppear = false }
+            if !fileState.hasAnyActiveGroup || isFirstAppear {
+                Task {
+                    try? await fileState.setToDefaultGroup()
                 }
             }
         }
     }
     
     @MainActor @ViewBuilder
-    private func content() -> some View {
+    private func contentView() -> some View {
         if #available(macOS 13.0, *), appPreference.sidebarLayout == .sidebar {
             ContentViewModern()
         } else {
@@ -190,200 +150,73 @@ struct ContentView: View {
         }
         .padding(40)
     }
-}
-
-struct PrintModifier: ViewModifier {
-    @Environment(\.alertToast) var alertToast
-
-    @EnvironmentObject var exportState: ExportState
-
-#if canImport(AppKit)
-    @State private var window: NSWindow?
-#elseif canImport(UIKit)
-    @State private var window: UIWindow?
-#endif
     
-    @State private var isPreparingForPrint = false
-    
-    func body(content: Content) -> some View {
-        content
-            .sheet(isPresented: .constant(isPreparingForPrint)) {
-                ProgressView {
-                    Text(.localizable(.generalLoading))
+    private func handleImport(_ notification: Notification) {
+        guard let urls = notification.object as? [URL] else { return }
+        if window?.isKeyWindow == true {
+            Task.detached {
+                do {
+                    try await fileState.importFiles(urls)
+                } catch {
+                    await alertToast(error)
                 }
-                .padding(.horizontal, 40)
             }
-            .bindWindow($window)
-#if os(macOS)
-            .onReceive(NotificationCenter.default.publisher(for: .togglePrintModalSheet)) { _ in
-                if window?.isKeyWindow == true {
-                    isPreparingForPrint = true
-                    Task.detached(priority: .background) {
-                        do {
-                            let imageData = try await exportState.exportCurrentFileToImage(
-                                type: .png,
-                                embedScene: false,
-                                withBackground: true
-                            ).data
-                            await MainActor.run {
-                                if let image = NSImage(dataIgnoringOrientation: imageData) {
-                                    exportPDF(image: image)
-                                }
+        }
+    }
+    
+    private func handleOpenFromURLs(_ notification: Notification) {
+        if let urls = notification.object as? [URL] {
+            fileState.temporaryFiles.append(contentsOf: urls)
+            fileState.temporaryFiles = Array(Set(fileState.temporaryFiles))
+            if !fileState.isTemporaryGroupSelected || fileState.currentTemporaryFile == nil {
+                fileState.isTemporaryGroupSelected = true
+                fileState.currentTemporaryFile = fileState.temporaryFiles.first
+            }
+        }
+    }
+    
+    // Check if it is first launch by checking the files count.
+    private func prepare() async {
+        do {
+            let isEmpty = try viewContext.fetch(NSFetchRequest<File>(entityName: "File")).isEmpty
+            isFirstImporting = isEmpty
+            if isFirstImporting == true, try await CKContainer.default().accountStatus() != .available {
+                isFirstImporting = false
+                return
+            } else if !isEmpty {
+                try await fileState.mergeDefaultGroupAndTrashIfNeeded(context: viewContext)
+            }
+        } catch {
+            alertToast(error)
+        }
+        
+        self.cloudContainerEventChangeListener?.cancel()
+        self.cloudContainerEventChangeListener = NotificationCenter.default.publisher(
+            for: NSPersistentCloudKitContainer.eventChangedNotification
+        ).sink { notification in
+            if let userInfo = notification.userInfo {
+                if let event = userInfo["event"] as? NSPersistentCloudKitContainer.Event {
+                    // On macOS, event.type will be `setup` only when first launched.
+                    // On iOS, event.type will be `setup` every time it has been launched.
+                    print("NSPersistentCloudKitContainer.eventChangedNotification: \(event.type), succeeded: \(event.succeeded)")
+                    if event.type == .import, event.succeeded, isFirstImporting == true {
+                        isFirstImporting = false
+                        Task { @MainActor in
+                            do {
+                                try? await Task.sleep(nanoseconds: UInt64(2 * 1e+9))
+                                try await fileState.mergeDefaultGroupAndTrashIfNeeded(context: viewContext)
+                            } catch {
+                                alertToast(error)
                             }
-                        } catch {
-                            await alertToast(error)
                         }
-                        await MainActor.run {
-                            isPreparingForPrint = false
-                        }
+                        self.cloudContainerEventChangeListener?.cancel()
                     }
                 }
             }
-#endif
+        }
     }
 }
 
-@available(macOS 13.0, *)
-struct ContentViewModern: View {
-    @Environment(\.managedObjectContext) private var viewContext
-    @Environment(\.horizontalSizeClass) private var horizontalSizeClass
-    @Environment(\.alertToast) var alertToast
-    @EnvironmentObject var fileState: FileState
-    @EnvironmentObject var appPreference: AppPreference
-    @EnvironmentObject var layoutState: LayoutState
-        
-    @State private var columnVisibility: NavigationSplitViewVisibility = .automatic
-    @State private var isSettingsPresented = false
-    
-    var body: some View {
-        NavigationSplitView(columnVisibility: $columnVisibility) {
-            if #available(macOS 14.0, iOS 17.0, *) {
-#if os(macOS)
-                SidebarView()
-                    .toolbar(content: sidebarToolbar)
-                    .toolbar(removing: .sidebarToggle)
-#elseif os(iOS)
-                if horizontalSizeClass == .compact {
-                    SidebarView()
-                        .toolbar(content: sidebarToolbar)
-                        .toolbar(removing: .sidebarToggle)
-                } else {
-                    SidebarView()
-                        .toolbar(content: sidebarToolbar)
-                }
-#endif
-            } else {
-                SidebarView()
-                    .toolbar(content: sidebarToolbar)
-            }
-        } detail: {
-             ExcalidrawContainerView()
-                .modifier(ExcalidrawContainerToolbarContentModifier())
-                .opacity(fileState.isInCollaborationSpace ? 0 : 1)
-                .overlay {
-                    ExcalidrawCollabContainerView()
-                        .opacity(fileState.isInCollaborationSpace ? 1 : 0)
-                        .allowsHitTesting(fileState.isInCollaborationSpace)
-                }
-        }
-#if os(macOS)
-        .removeSettingsSidebarToggle()
-#elseif os(iOS)
-        .sheet(isPresented: $isSettingsPresented) {
-            SettingsView()
-        }
-#endif
-        .onChange(of: columnVisibility) { newValue in
-            layoutState.isSidebarPresented = newValue != .detailOnly
-        }
-    }
-    
-    @ToolbarContentBuilder
-    private func sidebarToolbar() -> some ToolbarContent {
-        ToolbarItemGroup(placement: .primaryAction) {
-            // create
-            NewFileButton()
-        }
-        
-#if os(macOS)
-        // in macOS 14.*, the horizontalSizeClass is not `.regular`
-        // if horizontalSizeClass == .regular {
-            ToolbarItemGroup(placement: .destructiveAction) {
-                SidebarToggle(columnVisibility: $columnVisibility)
-            }
-        // }
-#elseif os(iOS)
-        ToolbarItemGroup(placement: .topBarLeading) {
-            Button {
-                isSettingsPresented.toggle()
-            } label: {
-                Label(.localizable(.settingsName), systemSymbol: .gear)
-            }
-        }
-#endif
-//        ToolbarItemGroup(placement: .confirmationAction) {
-//            Color.blue.frame(width: 10, height: 10)
-//        }
-//        ToolbarItemGroup(placement: .status) {
-//            Color.yellow.frame(width: 10, height: 10)
-//        }
-//        ToolbarItemGroup(placement: .principal) {
-//            Color.green.frame(width: 10, height: 10)
-//        }
-//
-//        ToolbarItemGroup(placement: .cancellationAction) {
-//            Color.red.frame(width: 10, height: 10)
-//        }
-//
-#if os(macOS)
-        /// It is neccessary for macOS to `space-between` the new button and sidebar toggle.
-        ToolbarItemGroup(placement: .secondaryAction) {
-            Color.clear
-        }
-#endif
-    }
-}
-
-struct ContentViewLagacy: View {
-    @Environment(\.alertToast) var alertToast
-    @EnvironmentObject var fileState: FileState
-    @EnvironmentObject var layoutState: LayoutState
-    
-    
-    var body: some View {
-        ZStack {
-            ExcalidrawContainerView()
-                .layoutPriority(1)
-                .modifier(ExcalidrawContainerToolbarContentModifier())
-                .opacity(fileState.isInCollaborationSpace ? 0 : 1)
-                .overlay {
-                    ExcalidrawCollabContainerView()
-                        .opacity(fileState.isInCollaborationSpace ? 1 : 0)
-                        .allowsHitTesting(fileState.isInCollaborationSpace)
-                }
-            
-            HStack {
-                if layoutState.isSidebarPresented {
-                    SidebarView()
-                        .frame(width: 374)
-                        .clipShape(RoundedRectangle(cornerRadius: 12))
-                        .background {
-                            RoundedRectangle(cornerRadius: 12)
-                                .fill(.regularMaterial)
-                                .shadow(radius: 4)
-                        }
-                        .transition(.move(edge: .leading))
-                }
-                Spacer()
-            }
-            .animation(.easeOut, value: layoutState.isSidebarPresented)
-            .animation(.easeOut, value: layoutState.isInspectorPresented)
-            .padding(.top, 10)
-            .padding(.horizontal, 10)
-            .padding(.bottom, 40)
-        }
-    }
-}
 
 #if DEBUG
 //struct ContentView_Previews: PreviewProvider {
