+++ ExcalidrawZ/Subscription/components/SwitchAppStoreVersionView.swift
@@ -1,62 +1,54 @@
 //
-//  MigrateToNewVersion.swift
+//  SwitchAppStoreVersionView.swift
 //  ExcalidrawZ
 //
-//  Created by Dove Zachary on 2024/8/14.
+//  Created by Dove Zachary on 4/2/25.
 //
 
 import SwiftUI
 
-import ChocofordEssentials
 import ChocofordUI
+import SwiftyAlert
 
-struct MigrateToNewVersionSheetViewModifier: ViewModifier {
-    @Binding var showMigrateSheet: Bool
-    
-    init(isPresented: Binding<Bool>) {
-        self._showMigrateSheet = isPresented
-    }
+struct SwitchAppStoreVersionViewViewModifier: ViewModifier {
+    @Binding var isPresented: Bool
     
     func body(content: Content) -> some View {
         content
-            .sheet(isPresented: $showMigrateSheet) {
-                MigrateToNewVersionSheetView()
-            }
-            .onAppear {
-                if !appVersion.contains("alpha"),
-                    !appVersion.contains("beta"),
-                   Bundle.main.bundleIdentifier == "com.chocoford.ExcalidrawZ" || Bundle.main.bundleIdentifier == "com.chocoford.ExcalidrawZ-Debug" {
-                    showMigrateSheet = !UserDefaults.standard.bool(forKey: "PreventShowMigrationSheet")
-                }
+            .sheet(isPresented: $isPresented) {
+                SwitchAppStoreVersionView()
+                    .padding(40)
+                    .frame(width: 660, height: 540)
+                    .swiftyAlert()
             }
     }
 }
 
-fileprivate struct MigrateToNewVersionSheetView: View {
+struct SwitchAppStoreVersionView: View {
     @Environment(\.managedObjectContext) private var viewContext
-    @Environment(\.dismiss) var dismiss
-    @State private var window: NSWindow?
-    
-    @AppStorage("PreventShowMigrationSheet") var notShowAgain = false
-    
+    @Environment(\.dismiss) private var dismiss
+    @Environment(\.alertToast) private var alertToast
     @State private var didArchive: Bool = false
-    
+
     var body: some View {
-        VStack(spacing: 20) {
-            Text(.localizable(.migrationSheetTitle))
-                .font(.largeTitle)
+        VStack {
+            Text(.localizable(.paywallSwitchToAppStoreDialogTitle))
+                .padding(.horizontal)
+                .font(.title)
+                .multilineTextAlignment(.center)
             
-            Text(.localizable(.migrationSheetBody))
+            Divider()
             
             VStack {
                 Text("❗️❗️❗️\(String(localizable: .migrationAttentionTitle))❗️❗️❗️").font(.headline)
                 Text(.localizable(.migrationAttentionContent))
+                    .multilineTextAlignment(.center)
             }
             .padding(.vertical)
             .padding(.horizontal, 40)
             .background {
                 RoundedRectangle(cornerRadius: 8)
-                    .fill(Color(hexString: "#f57c00"))
+                    .fill(Color(hexString: "#f57c00").opacity(0.6))
                 RoundedRectangle(cornerRadius: 8)
                     .stroke(Color(hexString: "#ffb74d"))
             }
@@ -64,6 +56,7 @@ fileprivate struct MigrateToNewVersionSheetView: View {
             GeometryReader { geometry in
                 let spacing: CGFloat = 10
                 HStack(spacing: spacing) {
+                    // Step 1
                     VStack(spacing: 20) {
                         Image(systemSymbol: .squareAndArrowUp)
                             .resizable()
@@ -77,7 +70,18 @@ fileprivate struct MigrateToNewVersionSheetView: View {
                         }
                         .frame(maxHeight: .infinity, alignment: .top)
                         AsyncButton { @MainActor in
-                            try archiveAllFiles(context: viewContext)
+                            do {
+                                try archiveAllFiles(context: viewContext)
+                            } catch {
+                                alertToast(
+                                    .init(
+                                        displayMode: .hud,
+                                        type: .regular,
+                                        title: String(localizable: .archiveDoneWithErrorAlertTitle),
+                                        subTitle: String(localizable: .archiveDoneWithErrorAlertSubtitle)
+                                    )
+                                )
+                            }
                             didArchive = true
                         } label: {
                             Text(.localizable(.migrationSheetButtonArchive))
@@ -85,6 +89,16 @@ fileprivate struct MigrateToNewVersionSheetView: View {
                     }
                     .padding()
                     .frame(width: (geometry.size.width - spacing) / 2, height: geometry.size.height)
+                    .overlay(alignment: .topLeading) {
+                        Image(systemSymbol: ._1Circle)
+                            .resizable()
+                            .scaledToFit()
+                            .frame(height: 100)
+                            .rotationEffect(.degrees(15))
+                            .offset(x: -30, y: -30)
+                            .foregroundStyle(.quaternary)
+                    }
+                    .clipShape(RoundedRectangle(cornerRadius: 12))
                     .background {
                         if #available(macOS 14.0, *) {
                             RoundedRectangle(cornerRadius: 12)
@@ -101,33 +115,43 @@ fileprivate struct MigrateToNewVersionSheetView: View {
                         }
                     }
                     
+                    // Step 2
                     VStack(spacing: 20) {
                         Image(systemSymbol: .squareAndArrowDown)
                             .resizable()
                             .scaledToFit()
                             .frame(height: 30)
                         VStack(spacing: 10) {
-                            Text(.localizable(.migrationSheetDownloadHeadline))
+                            Text(.localizable(.paywallSwitchToAppStoreStep2Title))
                                 .font(.headline)
                             
-                            Text(.localizable(.migrationSheetDownloadDescription))
+                            Text(.localizable(.paywallSwitchToAppStoreStep2Description))
                                 .foregroundStyle(.secondary)
                         }
                         .frame(maxHeight: .infinity, alignment: .top)
                         HStack {
-                            Link(destination: URL(string: "https://excalidrawz.chocoford.com")!) {
-                                Text("Download")
+                            Link(destination: URL(string: "https://apps.apple.com/app/excalidrawz/id6636493997")!) {
+                                Text(.localizable(.generalButtonDownload))
                             }
                             .disabled(!didArchive)
                             .if(!didArchive) { content in
                                 content
                                     .popoverHelp(.localizable(.migrationDownloadTooltip))
                             }
-                                
                         }
                     }
                     .padding()
                     .frame(width: (geometry.size.width - spacing) / 2, height: geometry.size.height)
+                    .overlay(alignment: .topLeading) {
+                        Image(systemSymbol: ._2Circle)
+                            .resizable()
+                            .scaledToFit()
+                            .frame(height: 100)
+                            .rotationEffect(.degrees(15))
+                            .offset(x: -30, y: -30)
+                            .foregroundStyle(.quaternary)
+                    }
+                    .clipShape(RoundedRectangle(cornerRadius: 12))
                     .background {
                         ZStack {
                             if #available(macOS 14.0, *) {
@@ -149,55 +173,17 @@ fileprivate struct MigrateToNewVersionSheetView: View {
                 .buttonStyle(.borderedProminent)
                 .controlSize(.large)
             }
-            Text(.localizable(.migrationSheetTips))
-                .padding(.horizontal, 40)
-              
-            HStack {
-                Toggle(isOn: $notShowAgain) {
-                    Text(.localizable(.migrationSheetButtonNeverShow))
-                }
-                .opacity(0)
-                Spacer()
-                Button {
-                    dismiss()
-                } label: {
-                    Text(.localizable(.migrationSheetButtonClose))
-                }
-                .controlSize(.large)
-                .buttonStyle(.borderless)
-                Spacer()
-                Toggle(isOn: $notShowAgain) {
-                    Text(.localizable(.migrationSheetButtonNeverShow))
-                }
+            
+            Button {
+                dismiss()
+            } label: {
+                Text(.localizable(.generalButtonClose))
             }
-            .padding(.horizontal, 20)
-        }
-        .multilineTextAlignment(.center)
-        .padding()
-        .frame(width: 600, height: 500)
-        .modifier(MigrateToNewVersionSheetBackgroundModifier())
-        .bindWindow($window)
-        .onAppear {
-            window?.backgroundColor = .clear
+            .buttonStyle(.borderless)
         }
     }
 }
-
-struct MigrateToNewVersionSheetBackgroundModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        if #available(macOS 13.0, *) {
-            content
-                .background(Color.accentColor.gradient)
-                .preferredColorScheme(.dark)
-        } else {
-            content
-                .background(Color.accentColor)
-                .preferredColorScheme(.dark)
-        }
-    }
-}
-
 
 #Preview {
-    MigrateToNewVersionSheetView()
+    SwitchAppStoreVersionView()
 }
